+++ test/test_unique_numbers.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'helper'
-
-class TestUniqueNumbers < Test::Unit::TestCase
-  def test_generates_unique_numbers
-    assert_nothing_raised { FFaker::Number.unique.number(digits: 1) }
-  end
-end